cart-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
